+++ CHANGES.md
@@ -26,6 +26,7 @@
 - Move functionality from implementation provider to definition provider in magik-language-server.
 - Implementation provider now provides impementations of abstract methods.
 - Add SelectionRangeProvider to magik-language-server.
+- Drop templated check support, including checks CommentRegularExpressionCheck and XPathCheck.
 - Various small fixes.
 
 0.7.1 (2023-02-21)